Simplified the example
