--- examples/advanced/streaming/jobs/dict_streaming/app/custom/defs.py
@@ -1,7 +1,0 @@
-STREAM_CHANNEL = "file_stream"
-TOPIC_INITIAL_FILE = "initial_file"
-TOPIC_ECHO_FILE = "echo_file"
-
-FILE_RTR_REQUEST_TOPIC = "rtr_file"
-FILE_RTR_STREAM_CHANNEL = "rtr_file_stream"
-FILE_RTR_STREAM_TOPIC = "rtr_file_stream"
