+++ nvflare/fuel/utils/fobs/__init__.py
@@ -16,8 +16,6 @@ from nvflare.fuel.utils.fobs.fobs import (
     auto_register_enum_types,
     deserialize,
     deserialize_stream,
-    get_class_name,
-    load_class,
     num_decomposers,
     register,
     register_data_classes,
